+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,30 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import './files/c.js';
+
+const random = Math.random();
+
+const { default: unknownObject } = random > 0.5 ? await import('./files/a.json', { assert: { type: 'json' } }) : await import('./files/b.json', { assert: { type: 'json' } })
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export {
+    unknownObject,
+    createMyServer,
+};
+
